se aplican modificaciones y nuevas validaciones

+++ src/main/java/com/centrosalud/centrosalud/controller/PacienteController.java
@@ -8,8 +8,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.centrosalud.centrosalud.model.ConsultaMedica;
+import com.centrosalud.centrosalud.model.Especialidad;
 import com.centrosalud.centrosalud.model.Paciente;
+import com.centrosalud.centrosalud.model.Profesional;
+import com.centrosalud.centrosalud.service.ConsultaMedicaService;
+import com.centrosalud.centrosalud.service.EspecialidadService;
 import com.centrosalud.centrosalud.service.PacienteService;
+import com.centrosalud.centrosalud.service.ProfesionalService;
+
+import io.micrometer.core.ipc.http.HttpSender.Response;
 
 import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -19,9 +27,11 @@ import java.util.List;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PutMapping;
-
+import org.apache.logging.log4j.spi.ObjectThreadContextMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.web.bind.annotation.RequestParam;
+
 
 
 
@@ -31,20 +41,123 @@ import org.slf4j.LoggerFactory;
 public class PacienteController {
 
     private static final Logger log = LoggerFactory.getLogger(PacienteController.class);
-
-    
     
     
     @Autowired
     private PacienteService pacienteService;
 
+    @Autowired
+    private ProfesionalService profesionalService;
+
+    @Autowired
+    private EspecialidadService especialidadService;
+
+    @Autowired
+    private ConsultaMedicaService consultaMedicaService;
+
     @GetMapping
     public List<Paciente> getAllPacientes() {
-        log.info("GET /students");
+        log.info("GET /pacientes");
         log.info("Retornando todos los pacientes");
         return pacienteService.getAllPacientes();
     }
+
+    @GetMapping("/especialidades")
+    public List<Especialidad> getAllEspecialidades() {
+        log.info("GET /especialidades");
+        log.info("Retornando todas las especialidades");
+        return especialidadService.getAllEspecialidad();
+    }
     
+    @PostMapping("/especialidades")
+    public ResponseEntity<Object> createEspecialidad(@RequestBody Especialidad especialidad) {
+        List<Especialidad> listaespecial = especialidadService.getAllEspecialidad();
+        for(Especialidad e : listaespecial)
+        {
+            if(e.getNombre().equals(especialidad.getNombre())){
+                log.error("La especialidad ya existe {}", especialidad.getNombre());
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("No se puede crear la especialidad " + especialidad.getNombre()));   
+            }
+        }
+        return ResponseEntity.ok(especialidadService.createEspecialidad(especialidad));
+    }
+
+
+    @GetMapping("/profesionales")
+    public List<Profesional> getAllProfesionales() {
+        log.info("GET /profesionales");
+        log.info("Retornando todos los profeisonales");
+        return profesionalService.getAllProfesionales();
+    }
+    
+    
+    @PostMapping("/profesionales")
+    public ResponseEntity<Object> createProfesional(@RequestBody Profesional profesional) 
+    {
+        List<Profesional> listaprofesional = profesionalService.getAllProfesionales();
+        for(Profesional p : listaprofesional)
+        {
+            if(p.getRut().equals(profesional.getRut()))
+            {
+                log.error("El profesional ya existe ID {}", profesional);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("El profesional ya existe " + profesional));
+
+            }
+
+        }
+
+        return ResponseEntity.ok(profesionalService.createProfesional(profesional));
+    }
+
+    @GetMapping("/consultas")
+    public List<ConsultaMedica> getAllConsultaMedicas() {
+        log.info("GET /consultas medicas");
+        log.info("Retornando todas las consultas medicas");
+        return consultaMedicaService.getAllConsultaMedica();
+    }
+
+    @PostMapping("/consultas")
+    public ResponseEntity<Object> createConsultas(@RequestBody ConsultaMedica consultaMedica) 
+    {
+        List<Profesional> listaprofesional = profesionalService.getAllProfesionales();
+        List<Paciente> listapaciente = pacienteService.getAllPacientes();
+        boolean existeprof = false;
+        boolean existepac = false;
+        for(Profesional p : listaprofesional)
+        {
+            if(p.getRut().equals(consultaMedica.getProfesional().getRut()))
+            {
+                existeprof = true;
+            }
+        }
+
+        if(!existeprof)
+        {
+            log.error("El profesional no existe {}", consultaMedica.getProfesional());
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("El profesional no existe ID " + consultaMedica.getProfesional()));
+
+        }
+
+        for(Paciente pac : listapaciente)
+        {
+            if(pac.getRut().equals(consultaMedica.getPaciente().getRut()))
+            {
+                existepac = true;
+            }
+
+        }
+
+        if(!existepac)
+        {
+            log.error("El paciente no existe {}", consultaMedica.getPaciente());
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("El paciente no existe ID " + consultaMedica.getPaciente()));
+        }
+
+        return ResponseEntity.ok(consultaMedicaService.createConsultaMedica(consultaMedica));
+    }
+    
+    
+
 
     @GetMapping("/{id}")
     public ResponseEntity<Object> getPacienteById(@PathVariable Long id) {
@@ -76,12 +189,20 @@ public class PacienteController {
     }
 
     @DeleteMapping("/{id}")
-    public void deletePaciente(@PathVariable Long id)
+    public ResponseEntity<Object> deletePaciente(@PathVariable Long id)
     {
+        Optional<Paciente> pac = pacienteService.getPacienteById(id);
+        if(pac.isEmpty())
+        {
+            log.error("El paciente no existe", id);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorResponse("El usuario que quiere eliminar no existe " + id));
+        }
+        log.info("Se ha eliminado el paciente");
         pacienteService.deletePaciente(id);
+        return ResponseEntity.ok(pac);
 
     }
-    
+
 
     static class ErrorResponse {
         private final String message;
